adds discard gem, move deletion of repositories in a rake task to run in background worker [fixes SCI-2448]

+++ db/migrate/20180518113526_remove_repository_rows_and_columns_without_repository.rb
@@ -1,7 +1,9 @@
 class RemoveRepositoryRowsAndColumnsWithoutRepository < ActiveRecord::Migration[5.1]
   def up
-    repository_ids = Repository.select(:id)
-    RepositoryRow.where.not(repository_id: repository_ids).delete_all
-    RepositoryColumn.where.not(repository_id: repository_ids).delete_all
+    if column_exists?(:repositories, :discarded_at)
+      repository_ids = Repository.select(:id)
+      RepositoryRow.where.not(repository_id: repository_ids).delete_all
+      RepositoryColumn.where.not(repository_id: repository_ids).delete_all
+    end
   end
 end
